+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - General
   - Update flake.nix to build with babashka. #1373
   - Bump clj-kondo to `2023.05.19-20230522.085211-4`.
+  
+- Editor
+  - New code actions: `Replace ':refer :all' with ':refer [my-refer]'` and `Replace ':refer :all' with alias`. #1575
 
 ## 2023.05.04-19.38.01